refactor: RacingcarManager.js 로직을 분리
- 기존의 RacingcarManager.js의 ui로직을 app.js로 이동
- 기존의 RacingcarManager.js의 로직을 domain 기능만 하도록 축소
